+++ bin/cli-helpers.js
@@ -1,5 +1,6 @@
 var path = require("path");
 var fsExtra = require("fs-extra");
+const { globSync } = require("glob");
 
 function loadVerifyExamplesConfig(configPath) {
   /* load the doc test config if we can find it
@@ -19,23 +20,16 @@ function loadVerifyExamplesConfig(configPath) {
     var elmJsonPath = findParentElmJson(path.dirname(configPath));
     elmJson = require(elmJsonPath);
   } catch (e) {
-    console.log(`Copying initial elm-verify-examples.json to ${configPath}`);
-    fsExtra.copySync(
-      path.resolve(__dirname, "./templates/elm-verify-examples.json"),
-      configPath
-    );
-
-    verifyExamples = require(path.resolve(
-      __dirname,
-      "templates/elm-verify-examples.json"
-    ));
+    var elmJsonPath = findParentElmJson(path.dirname(configPath));
+    elmJson = require(elmJsonPath);
+    verifyExamples = { tests: "all" };
   }
 
   return resolveTests(configPath, Object.assign({}, verifyExamples, elmJson));
 }
 
 function resolveTests(configPath, config) {
-  if (config.tests === "exposed") {
+  if (config.tests === "exposed" || config.tests.includes("exposed")) {
     if (config.type == "package") {
       config.tests = config["exposed-modules"].concat("./README.md");
     } else {
@@ -44,6 +38,17 @@ function resolveTests(configPath, config) {
       );
       process.exit(1);
     }
+  }
+  if (config.tests === "all" || config.tests.includes("all")) {
+    var allElmFiles = config["source-directories"]
+      .map((d) =>
+        globSync("**/*.elm", {
+          cwd: path.join(path.dirname(configPath), "..", d),
+        })
+      )
+      .flat()
+      .map(elmPathToModuleName);
+    config.tests = allElmFiles.concat("./README.md");
   }
   return config;
 }
@@ -57,6 +62,10 @@ function findParentElmJson(p) {
   } else {
     return findParentElmJson(path.dirname(p));
   }
+}
+
+function elmPathToModuleName(pathName) {
+  return pathName.slice(0, -4).replace(/\//g, ".");
 }
 
 module.exports = {